scsi: iscsi_tcp, libcxgbi: use init_cmd_priv/exit_cmd_priv
Implemet init_cmd_priv/exit_cmd_priv/cmd_size to have the block layer
allocate the iscsi structs for the driver

Note: Because for cxgbi we do not have access to the specific session we
are creating cmds for, all sessions get the max of what is on the host.
This is normally going to one so it should not be any different resource
wise.

Signed-off-by: Mike Christie <michael.christie@oracle.com>