scsi: libiscsi: add helper to complete a iscsi_task
We export functions to complete a PDU and a scsi task, but for mgmt tasks
some drivers have to fake the itt. Drivers like qedi, qla4xxx and
be2iscsi which don't have direct access to the PDU that was put on the
wire can now use this to complete the task directly.

Signed-off-by: Mike Christie <michael.christie@oracle.com>