vhost: move vhost worker creation to kick setup
The next patch will add a new ioctl that allows userspace to create
workers and bind them to devs and vqs after VHOST_SET_OWNER. To support
older tools, newer tools that want to go wild with worker threads, and
newer tools that want the old/default behavior this patch moves the
default worker setup to the kick setup.

After the first vq's kick/poll setup is done we could start to get works
queued, so this is the point when we must have a worker setup. If we are
using older tools or the newer tools just want the default single vhost
thread per dev behavior then it will automatically be done here. If the
tools are using the newer ioctls that have already created the workers
then we also detect that here and do nothing.

Signed-off-by: Mike Christie <michael.christie@oracle.com>
Reviewed-by: Stefan Hajnoczi <stefanha@redhat.com>