scsi: qedi: drop libiscsi itt use
qedi has several tags:

1. 16 bit hw/fw tag for each low level cq/sq entry.
2. libiscsi itt/tag which comes from libiscsi for each cmd/task.
3. 32 bit iscsi itt/tag used for each iscsi PDU. The driver combines 1
and 2 to form this.

This patch simplifies the driver so we only have #1. Note that in a future
patchset for qedi only we can go further and remove #1 and use the blk
tag. I'm not doing that in this set because the focus is on libiscsi and
iscsi_tcp.

Signed-off-by: Mike Christie <michael.christie@oracle.com>