scsi: libiscsi: remove ISCSI_TASK_ABRT_SESS_RECOV
The next patches simplfy the task state, so we can more easily use it to
detect if a command completed without the back lock because that is being
removed. This patch removes ISCSI_TASK_ABRT_SESS_RECOV since no one checks
for that specific state and only cares if completed one way or another.

Signed-off-by: Mike Christie <michael.christie@oracle.com>