scsi: be2iscsi: prep for back_lock removal
This patchset removes the back_lock. To prep for that this patch has
be2iscsi check if the cmd has completed in the driver under the wrb_lock
and then grab the task lock to check if it's completed in libiscsi. For
the case where we could get an abort for the cmd while we are processing
the completion we then get a ref to the task to make sure it doesn't
complete from under us while we process it.

Signed-off-by: Mike Christie <michael.christie@oracle.com>