scsi: qedi: Drop invalid completion checks
The next patches remove the SCp.ptr use in libiscsi, so this drops the
check in qedi. It is not replaced with a check for task->sc or the state
because it should never happen that it's already completed. If it did then
we are already accessing a freed/reused task, so none of the fields are
valid.

The patch also removes some invalid cmd checks in the same code path. If
we ever had a cmd with no request or request_queue then something went
wrong in scsi-ml or the block layer and we should have checks in there and
not in the driver.

Signed-off-by: Mike Christie <michael.christie@oracle.com>