scsi: libiscsi: update leadconn comments/locking
We don't need the frwd_lock to access the lead conn during queueing. It's
setup at creation time and then only cleared during destruction. During
conn destruction we've already blocked the session so queuecommand will
not be running, so it will never see a NULL leadconn pointer.

It was from when we supported MC/s before this code was merged upstream.
We've only supported the single conn session for 15 years now, so it seems
safe to start removing. This cleans up the non iscsi and SCSI EH paths. The
frwd lock is kind of intertwined with multiple checks in those paths and I
have a another patchset to fix up the tmf handling later.

Signed-off-by: Mike Christie <michael.christie@oracle.com>