scsi: libiscsi: Move suspend_tx check during queuecommand
Drivers that use the iSCSI host workqueue check ISCSI_SUSPEND_BIT when we
run iscsi_data_xmit so we don't need to check that bit in
iscsi_queuecommand for those drivers. This moves the check to the path
where we handle offload drivers that do not use the iSCSI host workqueue.

Signed-off-by: Mike Christie <michael.christie@oracle.com>