scsi: qedi: Drop back_lock use for cmd completions
The back_lock is not required for task/pdu completions or for gettting or
releasing a ref to a task. This removes the task completion uses in qedi.

Note: We must use the bh locking calls for the list_lock in the completion
paths. We could get a nop interval timeout which would result in libiscsi
calling into the driver to send a nop from softirq context. I also modified
the xmit paths to use the bh calls for the list_lock so it's easy to check
the lock use and perf did not change.

Signed-off-by: Mike Christie <michael.christie@oracle.com>