scsi: libiscsi: remove ISCSI_TASK_REQUEUE_SCSIQ
ISCSI_TASK_REQUEUE_SCSIQ is just kind of awkward and it's wrong. If in
queuecommand we are going to return SCSI_MLQUEUE_TARGET_BUSY then we
must have finished up cleaning up the cmd when queuecommand returns.
We can then never have a cmd in ISCSI_TASK_REQUEUE_SCSIQ state when we
are running iscsi_finish_task so that check is not needed. We also don't
need the refcount checks since we have to be freeing the cmd from this
code path.

This has us call the final cleanup code directly. Because the cmd was
allocated by the block/scsi layer and we've been under the frwd_lock the
entire time, we know that nothing else has a ref to the cmd and we can
just call the free function directly.

Signed-off-by: Mike Christie <michael.christie@oracle.com>