scsi: libiscsi: Rm frwd lock when checking session state
This drops the frwd lock requirement for the session state checks in
queuecommand. Like with the transport class case, we only needed it as a
hint since when the session is cleaned up we will block the session which
flushes the queues, and then we clean up all running IO. So the locking
just prevents cleaning up extra cmds. This patch is a prep patch and does
not help perf. This patch and the next ones are going to chip away the
reasons we need the frwd lock in the queuecommand code path for ihost
workq users like iscsi_tcp and cxgbi until there are none left and we can
remove it for those types of drivers.

Signed-off-by: Mike Christie <michael.christie@oracle.com>