kthread: allow caller to pass in user_struct
Currently, the kthreadd's user_struct has its processes checked against
the RLIMIT_NPROC limit. In cases like for vhost where the driver is making
a thread for userspace, then we want the userspace process to have its
processes count checked and incremented.

This patch allows the internal kthread code to take a user_struct and pass
it to copy_process. The next patches will then convert the fork/cred code
and allow kthread users to pass in the user_struct.

Signed-off-by: Mike Christie <michael.christie@oracle.com>